requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
boto3==1.34.133
botocore==1.34.133
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.5.4
Django==5.0.6
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.2
django-phonenumber-field==8.0.0
django-storages==1.14.3
django-timestampedmodel==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-excel==2.4.1
drf-spectacular==0.27.2
drf-writable-nested==0.7.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
factory-boy==3.3.0
Faker==25.9.2
flake8==7.1.0
flake8-docstrings==1.7.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
mccabe==0.7.0
numpy==2.0.0
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
phonenumbers==8.13.39
pillow==10.3.0
pluggy==1.5.0
prompt_toolkit==3.0.47
pycodestyle==2.12.0
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
referencing==0.35.1
rpds-py==0.18.1
s3transfer==0.10.2
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
